:truck: move: component 파일 분리 및 일괄 export 모듈 적용
