+++ grass-diary/src/pages/Diary/modal/UnmodifyModal.jsx
@@ -1,4 +1,4 @@
-import * as stylex from '@stylexjs/stylex';
+import stylex from '@stylexjs/stylex';
 
 const styles = stylex.create({
   background: {
