+++ grass-diary/src/pages/Intro/Intro.jsx
@@ -1,4 +1,4 @@
-import Header from '@components/Header';
+import { Header } from '@components';
 import {
   Container,
   Section,
