+++ grass-diary/src/pages/MyPage/Diary.jsx
@@ -6,8 +6,7 @@ import { useState } from 'react';
 
 import useUser from '@hooks/useUser';
 import useDiary from '@hooks/useDiary';
-import NormalLike from '@components/NormalLike';
-import MoodProfile from '@components/MoodProfile';
+import { NormalLike, MoodProfile } from '@components';
 
 const Pagination = ({ pageSize, onPageChange }) => {
   return (
