+++ grass-diary/src/pages/MyPage/MyPage.jsx
@@ -1,4 +1,4 @@
-import Header from '@components/Header';
+import { Header } from '@components';
 import { Container, MainContainer } from './myComponents';
 
 const MyPage = () => {
